Move code which handles internal trigger registration to st2common package.
Also explicitly call registration code inside the action runner and sensor
container instead of abusing module import side affects.
