+++ st2actions/st2actions/container/actionsensor.py
@@ -13,80 +13,21 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import httplib
-
-import eventlet
-import json
-import requests
-import requests.exceptions
-
 from oslo.config import cfg
 from st2common import log as logging
 from st2common.constants.triggers import INTERNAL_TRIGGER_TYPES
 from st2common.models.system.common import ResourceReference
 from st2common.transport.reactor import TriggerDispatcher
 
-ACTION_SENSOR_ENABLED = cfg.CONF.action_sensor.enable
-TRIGGER_TYPE_ENDPOINT = cfg.CONF.action_sensor.triggers_base_url
-TRIGGER_INSTANCE_ENDPOINT = cfg.CONF.action_sensor.webhook_sensor_base_url
-TIMEOUT = cfg.CONF.action_sensor.request_timeout
-MAX_ATTEMPTS = cfg.CONF.action_sensor.max_attempts
-RETRY_WAIT = cfg.CONF.action_sensor.retry_wait
-HTTP_POST_HEADER = {'content-type': 'application/json'}
+__all__ = [
+    'post_trigger'
+]
 
 LOG = logging.getLogger(__name__)
+
+ACTION_SENSOR_ENABLED = cfg.CONF.action_sensor.enable
 TRIGGER_DISPATCHER = TriggerDispatcher(LOG)
-
 ACTION_TRIGGER_TYPE = INTERNAL_TRIGGER_TYPES['action'][0]
-
-
-def _do_register_internal_trigger_types():
-    LOG.debug('Registering internal trigger types...')
-
-    for resource_name, trigger_definitions in INTERNAL_TRIGGER_TYPES.items():
-        for trigger_definition in trigger_definitions:
-            LOG.debug('Registering internal trigger: %s', trigger_definition['name'])
-            register_trigger_type(trigger_definition=trigger_definition, attempt_no=0)
-
-
-def register_trigger_type(trigger_definition, attempt_no=0):
-    LOG.debug('Attempt no %s to register trigger %s.', attempt_no, trigger_definition['name'])
-
-    payload = json.dumps(trigger_definition)
-
-    try:
-        r = requests.post(TRIGGER_TYPE_ENDPOINT,
-                          data=payload,
-                          headers=HTTP_POST_HEADER,
-                          timeout=TIMEOUT)
-        if r.status_code == httplib.CREATED:
-            LOG.info('Registered trigger %s.', trigger_definition['name'])
-        elif r.status_code == httplib.CONFLICT:
-            LOG.info('Trigger %s is already registered.', trigger_definition['name'])
-        else:
-            LOG.error('Seeing status code %s on an attempt to register trigger %s.',
-                      r.status_code, trigger_definition['name'])
-    except requests.exceptions.ConnectionError:
-        if attempt_no < MAX_ATTEMPTS:
-            retry_wait = RETRY_WAIT * (attempt_no + 1)
-            LOG.debug('    ConnectionError. Will retry in %ss.', retry_wait)
-            eventlet.spawn_after(retry_wait, register_trigger_type,
-                                 trigger_definition=trigger_definition,
-                                 attempt_no=(attempt_no + 1))
-        else:
-            LOG.warn('Failed to register trigger %s. Exceeded max attempts to register trigger.',
-                     trigger_definition['name'])
-    except:
-        LOG.exception('Failed to register trigger %s.', trigger_definition['name'])
-
-
-def register_internal_trigger_types():
-    if not ACTION_SENSOR_ENABLED:
-        return
-
-    # spawn a thread to process this in order to unblock the main thread which at this point could
-    # be in the middle of bootstraping the process.
-    eventlet.greenthread.spawn(_do_register_internal_trigger_types)
 
 
 def post_trigger(liveaction):
@@ -107,6 +48,3 @@ def post_trigger(liveaction):
         TRIGGER_DISPATCHER.dispatch(trigger, payload=payload)
     except:
         LOG.exception('Failed to fire trigger for liveaction %s.', str(liveaction.id))
-
-# TODO: This is awful, import shouldn't have side affects
-register_internal_trigger_types()
